pyproject.toml
[tool.flake8]
max-line-length = 88
ignore = ['E203']

[tool.pytest.ini_options]
pythonpath = [
    "src", "src/pokepapers"
]

[tool.poetry]
name = "pokepapers"
version = "0.1.0"
description = ""
authors = ["J. Trigueiros <jjtrigueiros@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
typer = "^0.7.0"
rich = "^13.3.3"
shellingham = "^1.5.0.post1"
pydantic = "^1.10.7"
pillow = "^9.5.0"
numpy = "^1.24.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"
pytest = "^7.2.2"
types-requests = "^2.28.11.17"
ruff = "^0.0.261"
pre-commit = "^3.2.2"
types-pillow = "^9.5.0.1"

[tool.poetry.scripts]
ppp = "pokepapers.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
